% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seqs.R
\name{add_seqs}
\alias{add_seqs}
\title{Add OTU sequences and their lengths to an OTU table}
\usage{
add_seqs(fasta, df, new_cols)
}
\arguments{
\item{fasta}{A character string specifying the path to/name of the OTU fasta file}

\item{df}{A data frame where each row corresponds to an OTU and includes a column named "OTU" to match names with the fasta file}

\item{new_cols}{A number specifying where the new columns should be inserted into df}
}
\value{
The original data frame with columns containing the DNA sequence (df$sequence) and its length (df$seq_length) for all OTUs
}
\description{
Extracts the DNA sequence and its length for each OTU from a fasta file and adds them to two columns of the OTU table data frame
}
\examples{
add_seqs(fasta = "~/Desktop/results/OTUs.fasta", df = COI_taxa, new_cols = 9)
}
