% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insect_merge.R
\name{lca_insect_merge}
\alias{lca_insect_merge}
\title{Merge LCA and insect taxonomic assignment data}
\usage{
lca_insect_merge(lca_taxa, taxdir, insect_path = "insect_classified.csv")
}
\arguments{
\item{lca_taxa}{A data frame where each row corresponds to an OTU with 10 or 11 columns including:
\itemize{
\item taxonomic information across 8 levels in descending order (domain, kingdom, phylum, class, order, family, genus, species)
\item OTU name (named "OTU")
\item information from blast (numberOfUnq_BlastHits, and pid if lca_progression was used)
}}

\item{taxdir}{The path to the directory to download (or that already contains) the NCBI taxonomy dump (THIS IS OVER 2 GB IN SIZE)}

\item{insect_path}{The path to the insect data file (default = "insect_classified.csv)}
}
\value{
A data frame containing taxonomic assignments from domain to species and information from either blast or insect (whichever assignment was kept) for all OTUs
}
\description{
Merges BLAST LCA and insect taxonomic assignment data for all OTUs into one data frame. If an OTU was assigned taxonomy by the LCA process (through BLAST), the insect and assignment related paratemers are nullified so the assignment method is clear. PLEASE NOTE THAT THE NCBI TAXONOMY DIRECTORY IS OVER 2GB IN SIZE. Make sure your computer has space for this and refer to this file path each time this function is used to save storage space.
}
\examples{
lca_insect_merge(lca_taxa = lca[,1:11], taxdir = "~/Desktop", insect_path = "~/Desktop/results/insect_results.csv")
}
