% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_progression.R
\name{lca_progression}
\alias{lca_progression}
\title{Merge multiple eDNAFlow LCA taxonomic assignment tables for more confident taxonomic assignments}
\usage{
lca_progression(glob_to_pid = "taxonomy_q[0-9]+_p", dir = getwd(), lulu = NA)
}
\arguments{
\item{glob_to_pid}{A character string that specifies the file naming convention for each OTU table with different lca parameters up to the percent identity (pid) value (which needs to be in the file name)}

\item{dir}{A character string that specifies the directory in which the lca taxonomy (identified in glob_to_pid) files are found (default = getwd())}

\item{lulu}{A boolean to specify whether you want lulu curated or non-lulu curated files if both can be found in the directory with the same glob_to_pid (default = NA)}
}
\value{
A data frame with one row for every OTU identified at any pid level with 11 columns including:
\itemize{
\item taxonomic information across 8 levels in descending order (domain, kingdom, phylum, class, order, family, genus, species)
\item OTU name (named "OTU")
\item the number of unique blast hits (numberOfUnq_BlastHits)
\item the highest percent identity level at which the OTU was classified (pid)
}
}
\description{
Merges a series of similarly named OTU tables generated from running the eDNAFlow LCA taxonomy assignment process on the same dataset with different minimum percent identity (pid) thresholds to keep only the highest level match per OTU. The main point of this is to provide the approximate pid match for each OTU that can be examined alongside the number of unique blast hits to determine whether there may be a dubious assignment from unclear, incomplete, or incorrect database listings.
}
\examples{
lca_progression()
lca_progression(glob_to_pid = "taxonomy_", dir = "~/Desktop/results", lulu = T)
}
