% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_report.R
\name{sample_report}
\alias{sample_report}
\title{Generate reports of reads and OTUs by sample}
\usage{
sample_report(
  df,
  first_sample_col,
  taxa_mat,
  min_reads_sample = 8,
  sep_human = T,
  sep_bacteria = F,
  sep_phyta = F,
  sep_eukaryota = F,
  sep_no_domain = F,
  custom_palette = F,
  report_graphs = T,
  report_text = T,
  return_list = F,
  file_prefix = NULL,
  outdir = getwd()
)
}
\arguments{
\item{df}{A data frame where each row corresponds to an OTU with columns containing at least:
\itemize{
\item taxonomic information across 8 levels (domain, kingdom, phylum, class, order, family, genus, species);
\item read counts per OTU from all samples sequenced (one column per sample) after all taxonomic and other data columns
}}

\item{first_sample_col}{A number denoting which column the sample read counts start (and continue to the end of the data frame)}

\item{taxa_mat}{A three-column matrix of character strings containing information about your taxa of interest (check if non-target taxa with reads you want reported have other sep_ options below):
\itemize{
\item column 1 - taxon name: like target_taxon in taxonomy_report()
\item column 2 - taxon rank: like taxon_rank in taxonomy_report()
\item column 3 - report name: what you want each group to be named (e.g., a common name, defaults to the taxon name in the first column if left blank)
\item Example:
taxa_mat = matrix(c(
"Metazoa", "kingdom", "Animal",
"Actinopteri", "class", "Fish",
"Viridiplantae", "kingdom", "Plant"),
byrow = T, ncol = 3)
}}

\item{min_reads_sample}{A number denoting the minimum quantity of OTU reads in a sample for it to be included in the generated GRAPHS (default = 8)}

\item{sep_human}{A boolean specifying whether human reads should be separated/tracked in their own category (default = T)}

\item{sep_bacteria}{A boolean specifying whether bacterial reads should be separated/tracked in their own category (default = F)}

\item{sep_phyta}{A boolean specifying whether eukaryotic algae/plant reads should be separated/tracked in their own category (default = F)}

\item{sep_eukaryota}{A boolean specifying whether non-target eukaryotic reads should be separated/tracked in their own category (default = F)}

\item{sep_no_domain}{A boolean specifying whether reads for which a domain could not be assigned should be separated/tracked in their own category (default = F)}

\item{custom_palette}{A character string containing a color palette to be interpreted by ggplot2 for created graphs (default = F)}

\item{report_graphs}{A boolean specifying whether or not to create graphs of the read and OTU counts of specified taxa across samples (default = T)}

\item{report_text}{A boolean specifying whether or not to create text reports of the read counts of specified taxa across samples (default = T)}

\item{return_list}{A boolean specifying whether or not to return the sample columns sorted into the groups as a list object is returned in R (default = F)}

\item{file_prefix}{A character string to be the prefix of the output file (default = NULL)}

\item{outdir}{A character string to specify the output directory of the output files (default = getwd())}
}
\value{
\itemize{
\item When report_graphs = T, graphs of overall and relative abundance of reads and OTUs per sample (grouped by target taxa specified in the taxa_mat) are returned.
\item When report_text = T, a text file with the overall and relative abundance of reads per sample is created (grouped by target taxa)
\item When return_list = T, a list containing two data frames is returned. The first is the sample read count columns with an additional "group" variable. The second object is a pivot_longer() version of this data frame grouped by the "group" variable.
}
}
\description{
Generates graphs, text reports, and data frames of samples sorted by pre-selected groups. This function will use only the 8 columns of taxonomic information (domain to species, in any location
across the data frame), as well as sample read counts starting from the first specified column of sample read counts (first_sample_col) through to the end of the data frame (so make sure your data frame only contains
columns of sample read counts after the specified first_sample_col).
\itemize{
\item The purpose of the "sep_..." group of arguments is to more easily sort out common contaminants or non-target sequence data instead of having to
put these in your taxa_mat. If you are interested in knowing how many reads were assigned to human, bacteria, land plants + algae, non-specified and non-target eukaryotes, or sequences that could not be to assigned
a domain, but one (or all) of these groups were not among your targeted taxa, it is better to use one of the "sep_" arguments. These arguments put reads and OTUs from these categories at the top of each sample's data
report, keeping targeted taxa at the bottom where counts and proportions can be more easily found and read. If a subgroup of one of these taxa for which a "sep_" argument exists is a target taxon (e.g. one phylum of
algae or genus of bacteria), the sep_ argument as TRUE will exclude target sequences labeling the remaining non-targets as "Other ..." (e.g., algae or bacteria).
\item If Primates (order) or Hominidae (family) are target taxa (in taxa_mat), the sep_human argument treats all sequences in the genus Homo as human. If neither of these two taxa are in the taxa_mat, all sequences
identified only to Primates or Hominidae are also considered human.
\item Custom color palettes must match the number of groups to be plotted, otherwise the default palette will be used. If you get warned that an incorrect number of colors were provided, check how many were plotted using
the default and then run the function again with the correct number of provided colors.
}
}
\examples{
sample_report(df = mifish_data, first_sample_col = 9, taxa_mat = matrix(c("Actinopteri", "class", "Fish"), nrow = 1))

taxa_groups = sample_report(df = coi_data, first_sample_col = 16,
                            taxa_mat = matrix(c("Metazoa", "kingdom", "Animal",
                                                "Arthropoda", "phylum", "Arthropod",
                                                "Rhodophyta", "phylum", "Red algae"),
                                              byrow = T, ncol = 3),
                             min_reads_sample = 2,
                             sep_bacteria = T,
                             sep_phyta = T,
                             sep_eukaryota = T,
                             sep_no_domain = T,
                             report_graphs = T,
                             custom_palette = ggthemes::colorblind_pal()(8),
                             report_text = F,
                             return_list = T,
                             file_prefix = "coi_samples",
                             outdir = "~/Desktop/results")
}
