% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_report.R
\name{taxonomy_report}
\alias{taxonomy_report}
\title{Generate a report of the taxonomic diversity detected within a target group}
\usage{
taxonomy_report(
  df,
  target_taxon,
  taxon_rank,
  dropped = "dropped",
  rm_humans = F,
  file_prefix = target_taxon,
  outdir = getwd()
)
}
\arguments{
\item{df}{A data frame where each row corresponds to an OTU with columns containing: taxonomic information across 8 levels in descending order}

\item{target_taxon}{A character string corresponding to the taxon for which a breakdown of nested taxa will be returned}

\item{taxon_rank}{A character string corresponding to the taxonomic level of the target taxon (one of the eight levels from the columns of df)}

\item{dropped}{A character string assigned to lower taxonomic levels that couldn't be resolved for an OTU (default = "dropped)}

\item{rm_humans}{A boolean to specify whether or not human OTUs should be included or removed (default = F)}

\item{file_prefix}{A character string specifying the prefix of the output file (default = target_taxon)}

\item{outdir}{A character string specifying the output directory of the text file (default = getwd())}
}
\value{
A text file listing the names of taxa nested within the target taxon, as well as the number of OTUs assigned to each of the nested taxa
}
\description{
Generates a list-style report containing lots of taxonomic information by OTU. The first section provides information about how many taxa were dropped at each level for
all OTUs/taxa. The next section does the same thing but only for OTUs in the target taxon. Finally, for each taxonomic level below the rank of the target taxon, a list of all taxa
found at this level and the number of OTUs that correspond to each taxon is provided. Unless primates or hominids are target taxa, this function treats all sequences identified as
Primates as human (in rm_humans).
}
\examples{
eDNA_taxon_report(df = df, target_taxon = "Metazoa", taxon_rank = "kingdom")
eDNA_taxon_report(df = coi_edna_data, file_prefix = "COI_eDNA_animal", target_taxon = "Metazoa",\
                  dropped = "unknown", rm_humans = T, outdir = "~/Desktop/results")
}
